chore(astro): move error-map
