+++ packages/astro/src/content/index.ts
@@ -1,5 +1,4 @@
 export { CONTENT_FLAG, PROPAGATED_ASSET_FLAG } from './consts.js';
-export { errorMap } from './error-map.js';
 export { attachContentServerListeners } from './server-listeners.js';
 export { createContentTypesGenerator } from './types-generator.js';
 export {
