+++ packages/astro/src/content/utils.ts
@@ -12,12 +12,9 @@ import type {
 	ContentEntryType,
 	DataEntryType,
 } from '../@types/astro.js';
-import { AstroError, AstroErrorData } from '../core/errors/index.js';
-
-import { MarkdownError } from '../core/errors/index.js';
+import { AstroError, AstroErrorData, MarkdownError, errorMap } from '../core/errors/index.js';
 import { isYAMLException } from '../core/errors/utils.js';
 import { CONTENT_FLAGS, CONTENT_TYPES_FILE, PROPAGATED_ASSET_FLAG } from './consts.js';
-import { errorMap } from './error-map.js';
 import { createImage } from './runtime-assets.js';
 
 /**
